esci-interpreter-perfection-v330: new download location, plugin registry change
The download paths on the EPSON servers are now such that I don't need
to fetch-restrict them any more. The downside? You can only get the
plugin binaries as part of a bundle tar file including the iscan and
iscan-data binaries as well.

I've moved the location of the plugin registry file from /var/lib/iscan
to /usr/$TARGET/lib/iscan to fix problems with cross.